refactor(update): 클린 코드
