+++ .gitignore
@@ -35,3 +35,6 @@ out/
 
 ### VS Code ###
 .vscode/
+
+application-test.yml
+/src/main/java/io/bareun/base/test/
