+++ src/main/java/io/bareun/base/api/request/ApiRequest.java
@@ -4,7 +4,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
-import org.springframework.util.MultiValueMap;
 
 import java.util.function.Consumer;
 
@@ -79,29 +78,6 @@ public interface ApiRequest<T> {
      */
     default Logger getLogger() {
         return LoggerFactory.getLogger(this.getClass());
-    }
-
-    /**
-     * 헤더에 키와 값을 추가하고 이 {@link ApiRequest}를 반환합니다.
-     *
-     * @param key 헤더 키
-     * @param value 헤더 값
-     * @return 헤더가 추가된 이 {@link ApiRequest}
-     */
-    default ApiRequest<T> header(String key, String value) {
-        getHeaders().add(key, value);
-        return this;
-    }
-
-    /**
-     * 주어진 헤더 맵을 추가하고 이 {@link ApiRequest}를 반환합니다.
-     *
-     * @param headers 추가할 헤더 맵
-     * @return 헤더가 추가된 이 {@link ApiRequest}
-     */
-    default ApiRequest<T> headers(MultiValueMap<String, String> headers) {
-        getHeaders().addAll(headers);
-        return this;
     }
 
     /**
